# Check buffer size to avoid copy in chunkreader