## Description
Choose whether to use internal buffer or use given buffer based on the size.

## Motivation and Context
This will allow us to remove the copy call, and improve performance since this
chuckreader is heavily used in minio.

## How to test this PR?
Existing unit tests

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Optimization (provides speedup with no functional changes)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation updated
- [ ] Unit tests added/updated