@kaplanlior 2021-05-17
Warp tool is used to simulate the stress test.
Average put throughput is increased by 4%.

Benchmark Results
----------------------------------------

**Before**
Operation: PUT
* Average: 198.71 MiB/s, 19.87 obj/s

Throughput, split into 297 x 1s:
 * Fastest: 245.7MiB/s, 24.57 obj/s
 * 50% Median: 198.3MiB/s, 19.83 obj/s
 * Slowest: 150.1MiB/s, 15.01 obj/s

**After**
Operation: PUT
* Average: 205.37 MiB/s, 20.54 obj/s

Throughput, split into 296 x 1s:
 * Fastest: 255.4MiB/s, 25.54 obj/s
 * 50% Median: 206.1MiB/s, 20.61 obj/s
 * Slowest: 157.6MiB/s, 15.76 obj/s