@kaplanlior 2021-05-26
## Before
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
cpu: DO-Premium-AMD
BenchmarkChunkReader
BenchmarkChunkReader/benchmarking_chunk_reader_10mb
BenchmarkChunkReader/benchmarking_chunk_reader_10mb-4                  1        5104228091 ns/op        1472968616 B/op16712158 allocs/op
BenchmarkChunkReader/benchmarking_chunk_reader_80mb
BenchmarkChunkReader/benchmarking_chunk_reader_80mb-4                  1        4915628543 ns/op        1472968480 B/op16712154 allocs/op
```
## After
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
cpu: DO-Premium-AMD
BenchmarkChunkReader
BenchmarkChunkReader/benchmarking_chunk_reader_10mb
BenchmarkChunkReader/benchmarking_chunk_reader_10mb-4                  1        4455736412 ns/op        1472968056 B/op16712151 allocs/op
BenchmarkChunkReader/benchmarking_chunk_reader_80mb
BenchmarkChunkReader/benchmarking_chunk_reader_80mb-4                  1        4433445763 ns/op        1472969008 B/op16712161 allocs/op
```